mesa/draw: fix -Wformat warning
fixes: 5791826b
cc @pepp @mareko

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/22605>
(cherry picked from commit a1098a1e45f4ad74d90261beb6e2594953f25540)

+++ src/mesa/main/draw.c
@@ -1641,7 +1641,7 @@ _mesa_validated_drawrangeelements(struct gl_context *ctx, GLenum mode,
          _mesa_warning(ctx, "Invalid indices offset 0x%" PRIxPTR
                             " (indices buffer size is %ld bytes)"
                             " or unallocated buffer (%u). Draw skipped.",
-                            start, index_bo->Size, !!index_bo->buffer);
+                            start, (long)index_bo->Size, !!index_bo->buffer);
          return;
       }
